deleted_later/requirements.txt
python-docx==0.8.11
streamlit==1.32.0
boto3==1.34.34
python-dotenv==1.0.1
gitpython==3.1.42
difflib==3.10.0
typer==0.9.0
rich==13.7.0 